Click event only when number loaded
If you click on the prev button before the first fetch returns, an error happens for the url is incorrect, because current_num is not initialized yet. The solution proposed here is to wait for the first fetch to complete before continuing adding events. Another solution could be to start with disabled buttons and then to activate them after the first fetch.